refactor: extract shared manifest/instruction helpers, de-duplicate
Two helper sets were copy-pasted across many patches:

- DOM (org.w3c.dom.Element) AndroidManifest helpers (childrenNamed,
  removeChildren, getOrCreateApplicationMetaData, disableComponents*, ...) were
  duplicated in 4 avito patches on top of the analytics copy.
- Instruction reference decoders (methodReferenceOrNull / fieldReferenceOrNull /
  stringReferenceOrNull) were duplicated across 7 avito/privacy/tbank patches.

Move both into a single app.shared package (ManifestXml.kt, InstructionHelpers.kt)
and import from there; drop the now-unused imports. The patcher provides no DOM
manifest helpers (only ResXmlElement-based) nor register-flow helpers, so these
stay the project's own — just de-duplicated. The per-patch registers()/
writesRegister() differ by opcode set and intentionally stay local.

Pure refactor, no behaviour change. Verified on 227.0: Block listings, Morphe
settings, Remove ads (manifest helpers) and UI tweaks (instruction helpers +
favorites fieldAccess) all apply with no failures; patched build boots clean on
a Pixel 3a.
