+++ patches/src/main/kotlin/app/avito/patches/blacklist/BlockListingsPatch.kt
@@ -8,6 +8,7 @@ import app.morphe.patcher.extensions.InstructionExtensions.instructionsOrNull
 import app.morphe.patcher.patch.PatchException
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patcher.patch.resourcePatch
+import app.shared.childrenNamed
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import org.w3c.dom.Element
@@ -15,16 +16,6 @@ import org.w3c.dom.Element
 private const val BLACKLIST_CLASS = "Lapp/avito/blacklist/Blacklist;"
 private const val BLOCK_MENU_CLASS = "Lapp/avito/morphe/MorpheBlockMenu;"
 private const val BLACKLIST_ACTIVITY = "app.avito.blacklist.BlacklistActivity"
-
-private fun Element.childrenNamed(name: String): List<Element> {
-    val nodes = childNodes
-    return buildList {
-        for (i in 0 until nodes.length) {
-            val node = nodes.item(i)
-            if (node is Element && node.nodeName == name) add(node)
-        }
-    }
-}
 
 /**
  * Registers the self-contained blacklist management screen
