+++ patches/src/main/kotlin/app/privacy/patches/analytics/DisableAppMetricaPatch.kt
@@ -4,6 +4,7 @@ import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patcher.patch.resourcePatch
 import org.w3c.dom.Element
+import app.shared.*
 
 private val disableAppMetricaManifestPatch = resourcePatch {
     execute {
